requirements.txt
amqp==2.6.1
asgiref==3.2.7
billiard==3.6.3.0
cachelib==0.1
celery==4.4.7
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Django==3.0.6
django-colorfield==0.3.2
gevent==20.6.2
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
persisting-theory==0.2.1
psycopg2-binary==2.8.5
pycparser==2.20
python-engineio==3.13.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zope.event==4.4
zope.interface==5.1.0
